Move Transaction type to own file
